requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.46.0
motor==3.7.1
pymongo==4.17.0
pydantic==2.13.3
pydantic-settings==2.14.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.26
pandas==3.0.2
numpy==2.4.4
scikit-learn==1.8.0
redis==7.4.0
python-dotenv==1.2.2
email-validator==2.3.0
pytest==8.4.2
httpx==0.28.1
